+++ .gitignore
@@ -9,6 +9,7 @@ _reports
 _UpgradeReport_Files/
 artifacts/
 Backup*/
+BenchmarkDotNet.Artifacts/
 bin
 Bin
 BuildOutput
