+++ Website.sln
@@ -9,6 +9,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.gitattributes = .gitattributes
 		.gitignore = .gitignore
 		.vsconfig = .vsconfig
+		benchmark.ps1 = benchmark.ps1
 		build.ps1 = build.ps1
 		Directory.Build.props = Directory.Build.props
 		Directory.Build.targets = Directory.Build.targets
@@ -38,8 +39,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".github", ".github", "{FBD3
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "workflows", "workflows", "{5B31CD2E-3F11-4609-88E8-117EF1389235}"
 	ProjectSection(SolutionItems) = preProject
+		.github\workflows\benchmark.yml = .github\workflows\benchmark.yml
 		.github\workflows\build.yml = .github\workflows\build.yml
 		.github\workflows\codeql-analysis.yml = .github\workflows\codeql-analysis.yml
+		.github\workflows\container-scan.yml = .github\workflows\container-scan.yml
 		.github\workflows\dependency-review.yml = .github\workflows\dependency-review.yml
 		.github\workflows\deploy.yml = .github\workflows\deploy.yml
 		.github\workflows\lighthouse.yml = .github\workflows\lighthouse.yml
@@ -56,6 +59,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".vscode", ".vscode", "{DF79
 		.vscode\launch.json = .vscode\launch.json
 		.vscode\tasks.json = .vscode\tasks.json
 	EndProjectSection
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "perf", "perf", "{A1F57AE9-C5B7-4503-8095-2C9376BAF873}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Website.Benchmarks", "perf\Website.Benchmarks\Website.Benchmarks.csproj", "{3548588A-27A1-47A7-BFB2-8D67D18EB81A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -79,6 +86,10 @@ Global
 		{26DB8EDD-BC86-428F-86A9-1C55711ABD0F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{26DB8EDD-BC86-428F-86A9-1C55711ABD0F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{26DB8EDD-BC86-428F-86A9-1C55711ABD0F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3548588A-27A1-47A7-BFB2-8D67D18EB81A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3548588A-27A1-47A7-BFB2-8D67D18EB81A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3548588A-27A1-47A7-BFB2-8D67D18EB81A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3548588A-27A1-47A7-BFB2-8D67D18EB81A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -91,6 +102,7 @@ Global
 		{E1197F63-A671-4484-B5BE-36C3561F4CBF} = {668E07A5-C40F-4694-9D3C-EBC1180656CF}
 		{26DB8EDD-BC86-428F-86A9-1C55711ABD0F} = {668E07A5-C40F-4694-9D3C-EBC1180656CF}
 		{DF799E65-EEEC-47DA-90E1-47A2965F132F} = {DD7EA67A-4FE1-4D4F-825E-51FF842107AB}
+		{3548588A-27A1-47A7-BFB2-8D67D18EB81A} = {A1F57AE9-C5B7-4503-8095-2C9376BAF873}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {15DDCA85-B2C3-4172-8203-DB0967289CA4}
